start adding pypy support to virtualenv.
The most important change so far is the way REQUIRED_MODULES are found:
before, virtualenv scanned /path/to/lib/python2.6/* looking for the
corresponding *.py files.  However, this does not work for pypy as there are
*three* directories where the stdlib resides (lib_pypy,
lib-python/modified-2.5.2 and lib-python/2.5.2), and they need to be scanned
in the correct order to ensure correctness.  The new approach is more direct:
simply try to import the module and let the interpreter to locate it, then use
its __file__ attribute to copy it to the virtualenv.

Moreover, site.py needs to be modified in order to add all that three
directories to sys.path.