Interpret path in filesystem encoding
(as opposed to platform default encoding)

This resolves pypa/virtualenv#186 ... but that leaves the virtualenv
creation to fail on the next hurdle, which is:

```
New python executable in FOO/bin/python
/opt/local/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/virtualenv-1.6.4-py2.7.egg/virtualenv.py:1233:
UnicodeWarning: Unicode unequal comparison failed to convert both
arguments to Unicode - interpreting them as being unequal
  if proc_stdout != os.path.normcase(os.path.abspath(home_dir)):
ERROR: The executable GED27/bin/python is not functioning
ERROR: It thinks sys.prefix is u'/Users/gthb/quickies/\xc6ttartal/GED27'
(should be '/Users/gthb/quickies/\xc3\x86ttartal/GED27')
ERROR: virtualenv is not compatible with this system or executable
```

and that, I guess is a Python bug. Still, this change gets virtualenv
doing the right thing, plus it makes the attempt fail in an
easier-to-understand manner.