Run easy_install script directly if not on Windows
i.e. only on Windows should it be invoked as an argument to the
python command (since it's necessary there); on other platforms it
should be invoked directly, in order to work around the python
interpreter borking with this error:

`Non-ASCII character '\xc3' in file [...] on line 1, but no encoding
declared; see http://www.python.org/peps/pep-0263.html for details`

if the path to the Python interpreter in the shebang line contains
non-ASCII characters.

Running such a python script directly works (on platforms that support
shebang in the first place), but passing it as an arg to the python
interpreter fails with the above error.

(Resolving the above error by inserting an encoding declaration doesn't
seem possible: inserting the encoding declaration as line 1 makes
shells unable to invoke the script directly --- they don't find the
shebang in line 1 --- and inserting it as line 2 doesn't make Python
accept the script as an arg to the python command.)

This changes nothing on Windows, but that ought to be OK, since
easy_install.py probably doesn't contain the shebang line anyway so
the problem doesn't come up there.