Merged in patch from: https://github.com/greghaskins/virtualenv/commit/d681c0ff9afbfba1fb225edc24fbab8bb9212678
To always create an activate on windows which can be used by cygwin and
mingw even for native windows python.

Added fix for bug #352844 "Windows: cannot create virtualenv on different
drive from system Python executable"

The real problem is that on windows, the DLL loading will look for the DLL
using the following criteria:

1. same directory as the exe
2. first occurance on the DLL lookup path in the registry
3. first occurance on the shell/system PATH vairable

The python.exe on windows relies on 1. This causes a problem if you have
multiple versions of python on your machine (32bit and 64bit for instance),
or if python was only installed for one user and thus is not placed in
registry path location (the windows system dir).

We should always copy the .dll as well on windows to make sure there is not
a conflict.