mageia multiarch fix for python virtualenv
If there exist /usr/bin/multiarch-platform, use it's output
as a help for defining the include path that is used
when installing python virtual env. If that executable
does not exist, define the include dir just like earlier.

This fixes the virtual python installation on mageia and
mandriva. (tested on with Mageia 2/X86_64)

Signed-off-by: Mika Laitio <mika.laitio@nokia.com>