pythonX.Y symlinks should be made on all platforms
The comment on line 1412 ("Some tools depend on pythonX.Y being
present") is accurate on all platforms, but the block of code that
ensures that pythonX.Y is present is inside a conditional that checks
"if '.framework' in prefix".

This commit moves the code out of the conditional so that it applies on
all platforms.