Fix long_description
Includes the right half of docs/index.rst (the one with links instead of the one with a `.. toctree:` directive).

Fixes the embarrassing unformatted PyPI page.  I'd appreciate a 'setup.py register', or a manual PyPI description update, because seeing unformatted PyPI
pages causes me pain in the aesthetic gland.