Don't use venv anymore on Debian-ized pythons (they broke venv with the `local/lib` and `dist-packages` stuff).
Things that are borken:

- bins install in `local/bin` instead of `bin`.
- packages go in `dist-packages` instead of `site-packages`.

Because of that it's disabled for pythons that seem to have those
patches. I made the checks so they probe sysconfig and distutils instead
of checking for distro names. This *should* make it future proof and
work as expected on Debian, Ubuntu 12 and 14, and the pythons from the
deadsnakes ppa.