Dereference symlinks more thoroughly in copyfile
copyfile was broken for symlinks that pointed to relative paths, and for
symlinks that pointed to further symlinks (which meant that virtualenv
failed to work in some cases when the Python executable it found was
a symlink, e.g. when Python had been installed using stow or homebrew;
issue #268).

Make it follow symlinks until it finds an actual file to copy.