Allow sourcing activate.sh in shells "exit on error" mode
The newly added `pydoc` unaliasing command ends with exit code 1 if `pydoc` alias is not set; this in turn means that sourcing `activate.sh` in a shell that had been set to exit on first error, upon sourcing that file, will exit on that line. And thanks to suppressing the output of `pydoc` there is no error message being displayed. Simply adding `|| true` should ensure that the command ends successfully.