Adapt for pip 10 changing location of main
This is really a special case hack. The only reason that I'm okay with importing from pip's internals here is because we're using pip to install pip and this is really the cleanest way to allow for the bootstrapping to work. Further, the script doesn't break any of pip's internal assumptions (like sole control of logging) so this is fine.