Symlink the python interpreter unless --always-copy is specified
If the standard library is symlinked into the virtual environment then
it makes sense to symlink the interpreter as well, so that the user is
not exposed to the risk of version skew between the python interpreter
and the standard library; nor the risk of the virtual environment
lacking security updates applied to the system python interpreter.

This requires Python 3.3, which adds support for the pyvenv.cfg file,
which, if present, suppresses the code that figures out the library
directories, etc., from the resolved path of the invoked executable.