activate.csh handle newlines -- don't use basename
Current `activate.csh` script is not robust against newlines in the prompt and uses `basename` unnecessarily.

For example, this `.tcshrc` file contains a prompt that breaks the script

```
onintr -

set tcshrc_nl = '\
'

 if ($?prompt) then
    set ellipsis = "..."
    set prompt = "$tcshrc_nl:q""%B[%.03]%b %# "
 endif

onintr
```

I modified the script to detect the presence of newlines and avoid changing the prompt at all in those cases.

The correct thing to do would probably be to add '[...]' to the beginning of the last line of the prompt, but I don't know of any easy way to accomplish that with csh. I briefly looked into `s///`-type substitution and using a temporary array variable to split the string into lines, but couldn't get it to work.

`tcsh` and `csh` also have the `:t` modifier for returning the last `/`-delimited component from a path, e.g.

```
[~] > set a = /a/b/c/d
[~] > echo $a:t
d
```