Eliminate spurious Azure fail; tweak clean_env init
It appears that for a completely non-colorized terminal fish's set_color
command returns either a null, or an empty string, or something.
This leads to incorrect string interpolation on the Azure test VMs.

See:
https://github.com/pypa/virtualenv/blob/94c6bcb4a3f8d7ff63e2215b18d76514d147a762/virtualenv_embedded/activate.fish#L92

The test in question fails, but the failing value of the prompt
is "env()". This implies that the interpolation is not "seeing anything"
from the output of the "(set_color normal)" invocation, and instead
putting the environment name ahead of, rather than into, the parentheses.

Hopefully this added env variable in the preamble of the script
will coerce fish into always including colorizing ANSI codes even
on the Azure VMs.

---

Also, tweak clean_env fixture to ensure VIRTUAL_ENV_DISABLE_PROMPT is
not present, as the tests all assume it is *unset* in the clean_env.

Also also, buff a number of comments.