Harmonize activate.xsh.
BEAR IN MIND, this only works for xonsh 0.8.9 and newer, and only
as long as xonsh doesn't change its implementation.

Probably the xonsh tests should be skipped on CI, since changes
outside the virtualenv codebase could cause breakage.

It would still be informative to have the xonsh tests in the
codebase, though -- perhaps a pytest CLI argument could be added
to enable them?