Fix error printing in bailout for Python < 2.7
Ironically, the code that tries to print an error message when
we're running on Python < 2.7 can only work on Python 2.7 or
higher! Prior to 2.7, the curly-braces style of string formatting
required all fields to be explicitly numbered; you have to use
{0}, you cannot use {}. So if you actually try to run virtualenv
on Python 2.6, you get an error when trying to print the error!
This should fix that, and print the right error...

Signed-off-by: Adam Williamson <awilliam@redhat.com>