Call towncrier via python instead of doing poor path reconstruction (#1848)
Trying to append an executable suffix to towncrier does not work
on Linux, as it can create absurd names like `towncrier.7` (`.7` coming
from `python3.7`).  Let's instead call it via `python -m towncrier`.
This should avoid relying on correct `PATH` (as I presume the original
code is meant to), while working correctly on all systems
and not requiring Python and towncrier to reside in the same directory.

Fixes #1847

Signed-off-by: Bernat Gabor <bgabor8@bloomberg.net>