Resolve symlinks when looking for the python interpreter
This makes the symlink point to the actual python exec, i.e. python3.6 instead of plain python3.
By doing this we ensure that the environment will survive a change in the default python3 version of the host system,
which can be caused by update-alternatives, an upgrade of the OS, etc.