Remove unnecessary special cases for bash/zsh
The internal shell utility hash is standard posix. There is no need to special-case it only for bash and zsh.
In fact, this check can break the script on shells such as the built-in sh on BSD systems, ksh, ash or 
busybox sh which are otherwise perfectly capable of running it. The script is a valid posix shell script, as
verified by shellcheck. The one feature that depends on bash (checking that the script is sourced rather 
than executed) will work only on bash, but it is only a safety check.

 https://pubs.opengroup.org/onlinepubs/9699919799/utilities/hash.html