Fix broken prompt set up by activate.bat (#2225)
Removing the stray closing parenthesis in activate.bat in commit
35ec441 broke the virtual environment prompt prefix in Windows cmd.
The reason for this was that the ill-formed closing parenthesis
somehow affected the variable expansion order.

This commit fixes the issue by not using a local ENV_PROMPT
variable to update the prompt.  (So an additional benefit of
this change is one less variable set in the environment.)

Fixes: https://github.com/pypa/virtualenv/issues/2225