When setting prefixes to empty strings, check the key names in addition to value
In Fedora, we are currently trying to come up with a safe approach that
will make `sudo pip` and similar install to /usr/local rather than /usr.

After years of only patching distutils for this,
we need to adapt our change to use sysconfig.

In the past, we've tried to use a custom installation scheme,
but that has yielded many unexpected side effects for our users.

Now we try to set the defaults of {base} and {platbase} to /usr/local instead.
However, such value is not in the prefixes listed in this check in virtualenv.

Here, I adapted the code to be more robust for such changes:

 - if the value is listed in prefixes, we still set to empty
 - if the key is listed in prefix_vars, we now also set to empty