Fix bad return code in bash activation if hashing is disabled
If hashing is disabled in bash, the `hash` command will return a nonzero
return code. Since it is the last command in the script that return code
will also be the "return code" of the source command, so if anyone uses

source activate.sh || die horribly

having hashing disabled will ruin their day.

Fix this by overriding the return code of `hash` if it's bad.
Add tests to verify a good return code and try with and without hashing.