feat: Write CACHEDIR.TAG file
Write a Cache Directory Tag file "CACHEDIR.TAG"
to created directory.

The CACHEDIR.TAG file is used by various tools to mark
a directory as cache, so that it can be handled differently.
Some backup tools look for this file to exclude the directory.
For example, tar includes the --exclude-caches option.

See https://bford.info/cachedir/ for more details.