Revert many parts of our wheel removal to support Python 3.8.
Pip 25.1 (which understands that wheel is not needed) is not supported
on Python 3.8, so we continue to ship pip 25.0.1 for Python 3.8, and
provide wheel.

This is a partial revert of commit db1d9fc02739d6388192eefb6befa9e4eccbe630.
Once support for Python 3.8 has been removed, this can be reverted.